Added a value to a mech's rd struct called xpmod, a float value.
This value, defaulting to 1.00 on template load, will allow for a per unit xp modification, which will allow for things like:

1) A player/team with crazy skills could have his xpmod tuned down to say, .5 - THey would get half xp for everything
2) A player/team with bad numbers on the field/bad mechs could have the xpmod tuned up to say, 1.5 - They would get roughly 50% more xp for everything

This does not (NOT!!) require a db rebuild.

Negative xpmod values result in a minimum of 1 XP

This only affects Gunnery attacks.

This can b set/retreived via btsetxcodevalue/btgetxcodevalue and @setxcode/@viewxcode

 7 files changed, 21 insertions(+), 12 deletions(-)



git-svn-id: svn://svn.code.sf.net/p/btonline-btech/code/btmux/trunk@640 e60b412a-5d77-420b-80c2-6cef64185626

+++ src/hcode/btech/p.btechstats.h
@@ -57,9 +57,9 @@ void AccumulateArtyXP(dbref pilot, MECH * attacker, MECH * wounded);
 void AccumulateComputerXP(dbref pilot, MECH * mech, int reason);
 int HasBoolAdvantage(dbref player, const char *name);
 void AccumulateGunXP(dbref pilot, MECH * attacker, MECH * wounded,
-    int numOccurences, int multiplier, int weapindx, int bth);
+    int numOccurences, float multiplier, int weapindx, int bth);
 void AccumulateGunXPold(dbref pilot, MECH * attacker, MECH * wounded,
-    int numOccurences, int multiplier, int weapindx, int bth);
+    int numOccurences, float multiplier, int weapindx, int bth);
 void fun_btgetcharvalue(char *buff, char **bufc, dbref player, dbref cause,
     char *fargs[], int nfargs, char *cargs[], int ncargs);
 void fun_btsetcharvalue(char *buff, char **bufc, dbref player, dbref cause,
